requirements.txt
boto3==1.26.73
botocore==1.29.73
jmespath==1.0.1
Pillow==9.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
